+++ voyageone-service2/voyageone-service2-dao/src/main/java/com/voyageone/service/daoext/cms/CmsBtJmPromotionTagProductDaoExt.java
@@ -12,7 +12,7 @@ import java.util.List;
 public interface CmsBtJmPromotionTagProductDaoExt {
     int deleteByCmsBtJmPromotionProductId(int cmsBtJmPromotionProductId);
 
-    int batchDeleteTag(@Param("listPromotionProductId") List<Long> listPromotionProductId);
+    int batchDeleteTag(@Param("listPromotionProductId") List<Integer> listPromotionProductId);
 
 
     int deleteByTagId(int tagId);
